Pull actual asset-creation functions out of manage_disaster.js (#285)
File was getting unwieldy. Only change (that I can remember) is making createScoreAsset (formerly run) return null instead of false if damageEnvelope was null, since that's what createScoreAsset returns when there's an error.
